requirements.txt
amqp==5.0.1
arabic-reshaper==2.1.0
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.3.0
boto3==1.14.56
botocore==1.17.56
celery==5.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1
django-common-helpers==0.9.2
django-cors-headers==3.5.0
django-cron==0.5.1
django-crontab==0.7.1
django-excel==0.0.10
django-extra-fields==3.0.0
django-filter==2.3.0
django-import-export==2.3.0
django-mailchimp-v1.3==1.4.0
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-rest-swagger==2.2.0
django-s3direct==1.1.8
django-storages==1.10
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-gis==0.16
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
docopt==0.4.0
docutils==0.15.2
drf-yasg==1.17.1
et-xmlfile==1.0.1
future==0.18.2
geocoder==1.38.1
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.2
html5lib==1.1
httpx==0.7.4
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
isort==5.4.2
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
lazy-object-proxy==1.4.3
lml==0.0.9
mailchimp==2.0.9
mailchimp-marketing==3.0.21
mailchimp3==3.0.14
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
msg91-otp==0.1.0
numpy==1.19.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==2.6.1
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
prompt-toolkit==3.0.7
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyexcel==0.6.4
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.9
pyexcel-xlsx==0.5.8
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-dotenv==0.19.2
python-http-client==3.3.1
pytz==2020.1
PyYAML==5.3.1
ratelim==0.1.6
razorpay==1.2.0
reportlab==3.5.53
requests==2.24.0
rfc3986==1.4.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sendgrid==3.6.5
sendgrid-django==4.2.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
tablib==2.0.0
texttable==1.6.3
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
vine==5.0.0
watchtower==1.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
# http://app.vehicleslabs.com.s3-website.ap-south-1.amazonaws.com/login